pyproject.toml
[project]
name = "grapRag"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.poetry.dependencies]
python = "^3.11.7"
langchain-groq = "^0.2.4"
langchain = "^0.3.18"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
langchain-huggingface = "^0.1.2"
matplotlib = "^3.10.0"
parsy = "^2.1"
ply = "^3.11"
streamlit = "^1.42.0"
instructor = "^1.7.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0","setuptools>=61.0"]
# build-backend = "poetry.core.masonry.api"
build-backend = "setuptools.build_meta"

